improve(CCTPFinalizer): Simulate transactions earlier to produce better failure logs
We can follow this pattern for the other finalizers too, I just wanted to try this for CCTP because there are some finalizations whose errors are only more clear when you simulate them directly before using the Multicall2 contract to simulate them
